refactored and removed types of reducer's states to general types file
