+++ frontend/src/constants/develop.ts
@@ -1,411 +1,1 @@
-import { locationType } from '../../types';
-
-export const positions: [number, number][] = [
-  [49.447731, 29.542721],
-  [50.447731, 29.542721],
-  [51.447731, 29.542721],
-
-  [49.447731, 30.542721],
-  [50.447731, 30.542721],
-  [51.447731, 30.542721],
-
-  [49.447731, 31.542721],
-  [50.447731, 31.542721],
-  [51.447731, 31.542721]
-];
-
-export const locations: locationType[] = [
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [49.447731, 29.542721],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [50.447731, 29.542721],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [51.447731, 29.542721],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [49.447731, 30.542721],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [50.447731, 30.542721],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [51.447731, 30.542721],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [49.447731, 31.542721],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [50.447731, 31.542721],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [51.447731, 31.542721],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [48.3416461723746, 22.917480468750004],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [49.49667452747045, 23.5546875],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [51.316880504045876, 24.279785156250004],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [51.316880504045876, 27.070312500000004],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [51.02757633780243, 30.124511718750004],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [48.25394114463434, 25.620117187500004],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [48.61838518688487, 27.685546875000004],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [47.517200697839414, 29.707031250000004],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [51.83577752045248, 33.59619140625001],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [50.55532498251967, 34.89257812500001],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [49.55372551347579, 37.48535156250001],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [49.210420445650314, 39.26513671875001],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [48.25394114463434, 39.15527343750001],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [47.15984001304432, 36.36474609375001],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [45.920587344733654, 29.531250000000004],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [46.437856895024204, 34.21142578125001],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  },
-  {
-    locationName: 'Synevyr',
-    rating: {
-      likes: 1,
-      dislikes: 0
-    },
-    coordinates: [45.72152152227956, 34.07958984375001],
-    arrayPhotos: [
-      'https://7chudes.in.ua/wp-content/uploads/2010/09/22829155.jpg'
-    ],
-    description:
-      "Озеро вважається найцікавішим об'єктом Національного природного парку «Синевир» і є однією з візитних карток Українських Карпат. Воно розташоване на висоті 989 метрів над рівнем моря, має середню площу 4—5 гектарів, його середня глибина становить 8—10 м, максимальна — 22-24 м. Через значну висоту над рівнем моря і порівняну глибоководність навіть у найтепліші дні прогріваються лише верхні 1—2 метри озера до максимальної температури 11—13°С.",
-    comments: []
-  }
-];
-
-// export const usersDataPlug: UserDataType[] = [
-//   {
-//     account: 'andrii@mapua.com',
-//     name: 'Andrii',
-//     userToken: '32qugc843cbu',
-//   },
-//   {
-//     account: 'roman@mapua.com',
-//     name: 'Roman',
-//     userToken: 'sservde324324',
-//   },
-//   {
-//     account: 'vlad@mapua.com',
-//     name: 'Vlad',
-//     userToken: '438pw9poln43dc3',
-//   },
-//   {
-//     account: 'yuriy@mapua.com',
-//     name: 'Yuriy',
-//     userToken: '234w3j0vpj5v',
-//   },
-//   {
-//     account: 'oleksandr@mapua.com',
-//     name: 'Oleksandr',
-//     userToken: 'kmndvxpin3v342',
-//   },
-// ];
+export {};
