--- frontend/src/redux/ts-types/locationList.tsx
@@ -1,9 +1,0 @@
-import { boundsType, latlngType } from '../../../types';
-
-export type locationsListStateType = {
-  locations: any[];
-  bounds: boundsType;
-  zoomPosition: latlngType;
-  locationName: string;
-  selectedFilters: string[];
-};
