--- frontend/src/redux/ts-types/mapInfo.ts
@@ -1,7 +1,0 @@
-import { boundsType } from '../../../types';
-
-export type mapInfoType = {
-  bounds: boundsType;
-  locationName: string;
-  selectedFilters: string[];
-};
