--- frontend/src/redux/ts-types/privateUserData.ts
@@ -1,1 +1,0 @@
-export {};
