--- frontend/src/redux/ts-types/user.tsx
@@ -1,5 +1,0 @@
-export type UserForm = {
-  displayName: string;
-  description: string;
-  imageUrl: FileList;
-};
