--- frontend/src/redux/ts-types/userAuth.tsx
@@ -1,5 +1,0 @@
-export type UserAuthState = {
-  loading: boolean;
-  error: {} | null;
-  isAuthorized: boolean;
-};
